backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.1
jinja2==3.1.4
PyJWT==2.9.0
eth-account==0.13.4
python-multipart==0.0.12
cryptography==43.0.3
langchain==0.3.7
langchain-weaviate==0.0.5
langchain-text-splitters==0.3.2
langchain-openai==0.2.8
weaviate-client==4.9.6
pypdf==5.1.0
pytest==8.3.3
